# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Mission Control API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /content:
    get:
      operationId: listContent
      tags: [content]
      summary: List content pipeline items
      parameters:
        - name: stage
          in: query
          schema:
            type: string
        - name: platform
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Content list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContentItem"
    post:
      operationId: createContent
      tags: [content]
      summary: Create content item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentItem"
  /content/{id}:
    get:
      operationId: getContent
      tags: [content]
      summary: Get single content item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Content item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentItem"
    patch:
      operationId: updateContent
      tags: [content]
      summary: Update content item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContentBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentItem"
    delete:
      operationId: deleteContent
      tags: [content]
      summary: Delete content item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /content/{id}/move:
    patch:
      operationId: moveContent
      tags: [content]
      summary: Move content item to a new stage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveContentBody"
      responses:
        "200":
          description: Moved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentItem"
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List calendar events
      parameters:
        - name: start
          in: query
          schema:
            type: string
        - name: end
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarEvent"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create calendar event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get single event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /agents:
    get:
      operationId: listAgents
      tags: [agents]
      summary: List all AI agents
      responses:
        "200":
          description: Agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
    post:
      operationId: createAgent
      tags: [agents]
      summary: Create an agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /agents/{id}:
    get:
      operationId: getAgent
      tags: [agents]
      summary: Get a single agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
    patch:
      operationId: updateAgent
      tags: [agents]
      summary: Update an agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgentBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"

  # ─── CONTACTS / CRM ───────────────────────────────────────────
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List all contacts
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create a contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      summary: Get single contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    patch:
      operationId: updateContact
      tags: [contacts]
      summary: Update contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    ContentItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        platform:
          type: string
          enum: [YouTube, X, LinkedIn, Instagram, Newsletter, Other]
        stage:
          type: string
          enum: [idea, scripting, recording, editing, review, scheduled, published]
        assignedDay:
          type: ["string", "null"]
        script:
          type: ["string", "null"]
        draftLink:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, platform, stage, createdAt, updatedAt]
    CreateContentBody:
      type: object
      properties:
        title:
          type: string
        platform:
          type: string
          enum: [YouTube, X, LinkedIn, Instagram, Newsletter, Other]
        stage:
          type: string
          enum: [idea, scripting, recording, editing, review, scheduled, published]
        assignedDay:
          type: ["string", "null"]
        script:
          type: ["string", "null"]
        draftLink:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required: [title, platform, stage]
    UpdateContentBody:
      type: object
      properties:
        title:
          type: string
        platform:
          type: string
          enum: [YouTube, X, LinkedIn, Instagram, Newsletter, Other]
        stage:
          type: string
          enum: [idea, scripting, recording, editing, review, scheduled, published]
        assignedDay:
          type: ["string", "null"]
        script:
          type: ["string", "null"]
        draftLink:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    MoveContentBody:
      type: object
      properties:
        stage:
          type: string
          enum: [idea, scripting, recording, editing, review, scheduled, published]
      required: [stage]
    CalendarEvent:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
          enum: [task, content, meeting, automation]
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        allDay:
          type: boolean
        createdAt:
          type: string
      required: [id, title, category, startDate, allDay, createdAt]
    CreateEventBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
          enum: [task, content, meeting, automation]
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        allDay:
          type: boolean
      required: [title, category, startDate, allDay]
...
    Agent:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        department:
          type: string
          enum: [Developers, Writers, Researchers, Operators]
        isLead:
          type: boolean
        status:
          type: string
          enum: [active, pending, error, idle]
        currentTask:
          type: ["string", "null"]
        lastActive:
          type: string
        responsibilities:
          type: ["string", "null"]
        avatarInitials:
          type: string
        tasksCompleted:
          type: integer
        successRate:
          type: integer
      required:
        - id
        - name
        - role
        - department
        - isLead
        - status
        - lastActive
        - avatarInitials
        - tasksCompleted
        - successRate
    CreateAgentBody:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        department:
          type: string
          enum: [Developers, Writers, Researchers, Operators]
        isLead:
          type: boolean
        responsibilities:
          type: ["string", "null"]
        avatarInitials:
          type: string
      required: [name, role, department, isLead, avatarInitials]
    UpdateAgentBody:
      type: object
      properties:
        status:
          type: string
          enum: [active, pending, error, idle]
        currentTask:
          type: ["string", "null"]
        responsibilities:
          type: ["string", "null"]
    Contact:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        handle:
          type: ["string", "null"]
        timezone:
          type: ["string", "null"]
        category:
          type: string
          enum: [Internal Team, Content Team, External, Clients]
        compensation:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, role, category, createdAt]
    CreateContactBody:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        handle:
          type: ["string", "null"]
        timezone:
          type: ["string", "null"]
        category:
          type: string
          enum: [Internal Team, Content Team, External, Clients]
        compensation:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required: [name, role, category]