# lib/api-zod/src/generated/types/createAgentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mission Control API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateAgentBodyDepartment } from "./createAgentBodyDepartment";

export interface CreateAgentBody {
  name: string;
  role: string;
  department: CreateAgentBodyDepartment;
  isLead: boolean;
  /** @nullable */
  responsibilities?: string | null;
  avatarInitials: string;
}