# lib/api-zod/src/generated/types/createContactBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mission Control API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateContactBodyCategory } from "./createContactBodyCategory";

export interface CreateContactBody {
  name: string;
  role: string;
  /** @nullable */
  handle?: string | null;
  /** @nullable */
  timezone?: string | null;
  category: CreateContactBodyCategory;
  /** @nullable */
  compensation?: string | null;
  /** @nullable */
  notes?: string | null;
}